Enable "Joins" e2e group
