+++ .circleci/config.yml
@@ -1379,6 +1379,7 @@ workflows:
                   "dashboard-filters",
                   "dashboard-filters-sql",
                   "downloads",
+                  "joins",
                   "models",
                   "moderation",
                   "native",
