+++ frontend/test/metabase/scenarios/joins/reproductions/13468-ui-overlay-after-save.cy.spec.js
