+++ frontend/test/metabase/scenarios/joins/reproductions/18589-numeric-binning-in-joins.cy.spec.js
