+++ frontend/test/metabase/scenarios/joins/reproductions/18630-field-literals-in-joins.cy.spec.js
