+++ frontend/test/metabase/scenarios/joins/reproductions/18818-crash-when-joining-on-custom-column.cy.spec.js
